requirements.txt
sqlalchemy>=1.4
psycopg2-binary>=2.9
python-dotenv>=0.21
